Implement the Composition API for the: - Axis; - Ticks; - CatesianChartGrid; - ScatterLineSeries, - BarSeries; - CandleStickSeries; - OhlcSeries; - Trend Line and Financial Indicator;
Added ContainerVisualsFactoruy class that provides the ability to choose whether the visual elements should be drawn using the Composition API, to change the ContainerVisual used for the drawing of the element (currently only SpriteVisual could be used) and to modify the ContainerVisual by changing its Offset, Size, Brush etc.

Implemented a LabelCreator interface tha provides the ability using a bool method to indicate whether a Label should be created or not - that provides the ability to choose the exact Labels the user could create based on some implemetned logic and could improve the performace a lot for specific scenaries where only a few specific labels needs to be visualized.