Fix order of operations bug in DataGrid SelectItem and SelectCell
When using extended selection mode, SelectItem and SelectCell would only briefly select the desired item before clearing the selection and leaving nothing selected. This change fixes the order of operations so that the selection is cleared first before making the new selection.

Fix for: https://github.com/telerik/UI-For-UWP/issues/462